Perbaikan Bug & Penambahan Multiple Account

+++ app/controllers/data_gurus_controller.ts
@@ -82,6 +82,33 @@ export default class DataGurusController {
     return inertia.render('Guru/Create', { session: session.flashMessages.all() })
   }
 
+  public async cekUser({ params, response }: HttpContext) {
+    const trx = await db.transaction()
+    const email = params.id
+    let user = await User.query({ client: trx })
+      .preload('dataGuru')
+      .preload('dataStaf')
+      .where('email', email)
+      .first()
+    let dataGuru = user?.dataGuru
+    if (dataGuru) {
+      response.json({
+        status: 'notReady',
+        message: 'Tidak Bisa Membuat Lebih 1 Data Dengan Guru Yang Sama',
+      })
+    } else if (user && dataGuru == null) {
+      response.json({
+        status: 'ready',
+        data: user,
+      })
+    } else {
+      response.json({
+        status: 'notReady',
+        message: 'Tidak ada data! Buat data Baru',
+      })
+    }
+  }
+
   public async store({ request, response, session }: HttpContext) {
     const trx = await db.transaction()
 
@@ -92,7 +119,16 @@ export default class DataGurusController {
         ? await this.uploadFile(request.file('guru.fileFoto'), String(payload.guru.nip))
         : null
 
-      const user = await User.create({ ...payload.user, role: 'Guru' }, { client: trx })
+      let user = await User.query({ client: trx }).where('email', payload.user.email).first()
+      if (user) {
+        user.role = 'Staf'
+        user.save()
+      }
+
+      if (!user) {
+        user = await User.create({ ...payload.user, role: 'Guru' }, { client: trx })
+      }
+
       await DataGuru.create(
         { ...(payload.guru as any), userId: user.id, fileFoto: fileFoto },
         { client: trx }
@@ -170,10 +206,16 @@ export default class DataGurusController {
     try {
       const { id } = params
       const guru = await DataGuru.query().where('nip', id).firstOrFail()
-      const user = await User.query().where('id', guru.userId).firstOrFail()
+      const user = await User.query()
+        .where('id', guru.userId)
+        .preload('dataGuru')
+        .preload('dataStaf')
+        .firstOrFail()
       await this.deleteFile(guru.fileFoto)
       await guru.delete()
-      await user.delete()
+      if (!user.dataStaf) {
+        await user.delete()
+      }
       session.flash({
         status: 'success',
         message: 'Data guru berhasil dihapus.',
